requirements.txt
pandas
streamlit
openpyxl
Pillow
matplotlib
seaborn
